requirements.txt
customtkinter>=5.2.2
paho-mqtt>=2.1.0
matplotlib>=3.8.0
cryptography>=42.0.0
roboticstoolbox-python>=1.1.0
spatialmath-python>=1.1.11
numpy>=1.26.0
